+++ HISTORY.rst
@@ -7,6 +7,7 @@ Historia zmian
 -----------
 
 * nie używaj tagów HTML w generowanych raportach 3N (b/n),
+* zawężaj raporty 3N do zakresu lat 2017-2021 (b/n),
 
 202201.1081
 -----------